[INDY-2154] Some refactorings
